Remove non-semantic visual code formatting
A number of text files in this repository use empty space characters in
places where they have no technical significance. This practice is
generally followed to vertically align information when text is rendered
visually. For users of assistive technologies, it may not be apparent
that this practice is being followed, and it may be burdensome to
maintain the convention.

Remove the formatting to promote equity in the project's contribution
experience.