Correct generated code (#655)
When the build script is run in Windows, strings which describe paths
will use the reverse solidus character to separate directory names and
file names. This character takes on special meaning within JavaScript
string literals (it begins an escape sequence), so it must be
transformed if it is to retain its intended purpose in that context.

Use `JSON.stringify` to properly sanitize a value which is inserted into
JavaScript program code.