Correct expression of non-blocking code
By applying the `await` operator, the script's invocations of the
Node.js filesystem methods indicate an intention to avoid blocking the
main thread of execution. However, both call sites use
non-promise-returning APIs leading to blocking behavior in the first
case (via the `unlinkSync` method) and non-deterministic behavior in the
second case (via the `appendFile` method).

Use the `fs/promises` version of Node.js's filesystem methods so that
the script is not blocked by I/O and so that it consistently reports
errors.