Change 'DescriptionListBoundary' assertion to 'listBoundary' Made the assertion: "List boundary is conveyed".
The description list boundary is technically more correct. However, the ARIA mapping is actually "association list" and every screen reader is handling this differently. There has been discussion in ARIA about this mapping. No one seems to be happy with it.

I am not confident it would be a good idea to fail screen readers that simply call it a list. That is what NVDA and Narrator do. JAWS still calls it a "definition list". VoiceOver calls it a description list.

Open to further discussion about this change.