Fixes assertion validation errors that were causing builds to hang (#1247)
* refactor assertion validation

and adds isolated unit tests

* adds docstrings

* update and simplify assertionsFormat

The previous version would not return anything or throw an error when the string tested contained a '?'. This updated version does return false and throw an error as expected.

* update assertionsFormat to include ":"

This more closely follows the examples in the wiki: https://github.com/w3c/aria-at/wiki/Test-Format-Definition-V2#assertions

* Generate .html source files with scripts automatically

* removes ? as valid character in assertion format

* fixup! lint

@howard-e @jugglinmike I'm wondering tho if we want to exclude the `.github` folder from the prettier linter?

* limits valid numeric chars in assertion ID's

* adds back priority tests in `assertionsFormat`