Tests 9 through 12: Clarify wording of titles
Distinguish the task being performed from the test condition. In test 9 through 12, the commands move to the next or previous tab without regard for whether the tab is selected. However, the conditions are different for each test. In test 9 and 10, the next tab is not selected. In test 11 and 12, the next tab is selected.

Example:

Changed "Navigate to the next not selected tab in a tab list"
to: "Navigate to the next tab in a tab list when the next tab is not selected"